+++ packages/@weresk/locales/src/index.ts
@@ -1,3 +1,4 @@
-// export * from "./lib";
-// export * from "./types";
-// export * from "./utils";
+export * from "./lib";
+export * from "./schemas";
+export * from "./types";
+export * from "./utils";