+++ packages/@weresk/locales/src/lib/index.ts
@@ -1,1 +1,1 @@
-// export * from "./language-filter";
+export * from "./language-filter";