+++ packages/@weresk/locales/src/lib/language-filter/index.ts
@@ -1,0 +1,1 @@
+export { default as languageFilterConfig } from "./languageFilterConfig";