+++ packages/@weresk/locales/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @weresk/locales
 
+## 0.0.2
+
+### Patch Changes
+
+- Adding schemas, types and utils to @weresk/locales
+
 ## 0.0.1
 
 ### Patch Changes