+++ packages/@weresk/maket/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @weresk/maket
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @weresk/core@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes